requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.3
Markdown==3.7
openai==1.51.2
pandas==2.2.3
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.1
reportlab==4.2.2
Requests==2.32.3
